Update TracingDriver.java
Current version doesn't allow you to pass Tracer instance to the TracingConnection.  

It only works with GlobalTracer.get, which doesn't work when you have multiple tracers in the same jvm.

These changes allow you to (1) set and pass a tracer and (2) to extend this class if you need different behavior.